added movie search by api

+++ app/src/main/java/com/dayker/viewed/watchedmovies/data/repository/WatchedMoviesRepositoryImpl.kt
@@ -12,7 +12,7 @@ class WatchedMoviesRepositoryImpl(
         return dao.getMovies()
     }
 
-    override suspend fun getMovieById(id: Int): Movie? {
+    override suspend fun getMovieById(id: Long): Movie? {
         return dao.getMovieById(id)
     }
