+++ app/src/main/java/com/dayker/viewed/watchedmovies/domain/repository/WatchedMoviesRepository.kt
@@ -6,7 +6,7 @@ import kotlinx.coroutines.flow.Flow
 interface WatchedMoviesRepository {
 
     fun getMovies(): Flow<List<Movie>>
-    suspend fun getMovieById(id: Int): Movie?
+    suspend fun getMovieById(id: Long): Movie?
     suspend fun insertMovie(movie: Movie): Long?
     suspend fun deleteMovie(movie: Movie)
 }
