+++ app/src/main/java/com/dayker/viewed/watchedmovies/domain/usecase/GetWatchedMovieUseCase.kt
@@ -5,5 +5,5 @@ import com.dayker.viewed.watchedmovies.domain.repository.WatchedMoviesRepository
 class GetWatchedMovieUseCase(
     private val repository: WatchedMoviesRepository
 ) {
-    suspend operator fun invoke(id: Int) = repository.getMovieById(id)
+    suspend operator fun invoke(id: Long) = repository.getMovieById(id)
 }
